updated to unsigned int

+++ 0x02-search_algorithms/0-linear.c
@@ -8,13 +8,13 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	unsigned int i;
 
 	if (array != NULL)
 	{
 		for (i = 0; i < size; i++)
 		{
-			printf("Value checked array[%u] = [%u]\n", i, array[i]);
+			printf("Value checked array[%d] = [%d]\n", i, array[i]);
 			if (array[i] == value)
 			{
 				return (i);
